@@ -1,10 +1,10 @@
 export default {
   transform: {},
+  extensionsToTreatAsEsm: ['.js'],
   moduleNameMapper: {
     '^(\\.{1,2}/.*)\\.js$': '$1',
   },
   testEnvironment: 'node',
-  testMatch: ['**/__tests__/**/*.js', '**/?(*.)+(spec|test).js'],
-  verbose: true,
-  moduleFileExtensions: ['js', 'json', 'node']
+  testMatch: ['**/__tests__/**/*.test.js'],
+  verbose: true
 }; 
